scaffolds/slack-bridge/pyproject.toml
[project]
name = "slack-bridge"
version = "0.1.0"
description = "Slack ↔ Claude Code bridge — Socket Mode listener that routes Slack messages to claude CLI"
requires-python = ">=3.11"
dependencies = [
    "slack-sdk>=3.27",
    "python-dotenv>=1.0",
]

[project.scripts]
slack-bridge = "slack_bridge.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slack_bridge"]
